Revert "merge setheum_currency to stp258_currency"
