+++ src/serp_tes.rs
@@ -1,4 +1,5 @@
 use codec::FullCodec;
+use crate::setheum_currency::{BalanceStatus, LockIdentifier};
 pub use frame_support::Parameter;
 use sp_runtime::{
 	traits::{
